push code save msg player

+++ Lottery.WebMvcCore/Lottery.WebMvc/Models/Calculation3Model.cs
@@ -7,8 +7,8 @@
         public int IdKhach { get; set; }
         public int Mien { get; set; }
         public DateTime HandlByDate { get; set; }
-        public double CachTrungDaThang { get; set; }
-        public double CachTrungDaXien { get; set; }
+        public int CachTrungDaThang { get; set; }
+        public int CachTrungDaXien { get; set; }
         public bool IsSave { get; set; }
         public string SynTax { get; set; }
     }
